+++ src/django_mysql/models/fields/tiny_integer.py
